Rename to redirect. Move channel id and message to variables.
